+++ src/commands/path/path.generate.ts
@@ -10,6 +10,7 @@ import { isArgo } from '../../utils/argo.js';
 import { slugify } from '../../utils/slugify.js';
 import { config, createTiff, registerCli, verbose } from '../common.js';
 import { dataCategories } from './path.constants.js';
+import { formatDate } from './path.date.js';
 
 export interface PathMetadata {
   targetBucketName: string;
@@ -94,10 +95,10 @@ export const commandGeneratePath = command({
 });
 
 /**
- *Generates target path based on dataset category.
+ * Generates target path based on dataset category.
  *
- * @param {PathMetadata} metadata
- * @returns {string}
+ * @param metadata
+ * @returns
  */
 export function generatePath(metadata: PathMetadata): string {
   const name = formatName(metadata.region, metadata.geographicDescription);
@@ -106,74 +107,40 @@ export function generatePath(metadata: PathMetadata): string {
   if (metadata.category === dataCategories.SCANNED_AERIAL_PHOTOS) {
     // nb: Historic Imagery is out of scope as survey number is not yet recorded in collection metadata
     throw new Error(`Automated target generation not implemented for historic imagery`);
-  } else if ([dataCategories.URBAN_AERIAL_PHOTOS, dataCategories.RURAL_AERIAL_PHOTOS].includes(metadata.category)) {
-    return `s3://${metadata.targetBucketName}/${metadata.region}/${surveyName}_${metadata.gsd}m/rgb/${metadata.epsg}/`;
-  } else if (metadata.category === dataCategories.SATELLITE_IMAGERY) {
-    return `s3://${metadata.targetBucketName}/${metadata.region}/${surveyName}_${metadata.gsd}m/rgb/${metadata.epsg}/`;
-  } else if ([dataCategories.DEM, dataCategories.DSM].includes(metadata.category)) {
-    return `s3://${metadata.targetBucketName}/${metadata.region}/${surveyName}/${metadata.category}_${metadata.gsd}m/${metadata.epsg}/`;
-  } else {
-    throw new Error(`Path Can't be generated from collection as no matching category: ${metadata.category}.`);
   }
+
+  if ([dataCategories.URBAN_AERIAL_PHOTOS, dataCategories.RURAL_AERIAL_PHOTOS].includes(metadata.category)) {
+    return `s3://${metadata.targetBucketName}/${metadata.region}/${surveyName}_${metadata.gsd}m/rgb/${metadata.epsg}/`;
+  }
+
+  if (metadata.category === dataCategories.SATELLITE_IMAGERY) {
+    return `s3://${metadata.targetBucketName}/${metadata.region}/${surveyName}_${metadata.gsd}m/rgb/${metadata.epsg}/`;
+  }
+
+  if ([dataCategories.DEM, dataCategories.DSM].includes(metadata.category)) {
+    return `s3://${metadata.targetBucketName}/${metadata.region}/${surveyName}/${metadata.category}_${metadata.gsd}m/${metadata.epsg}/`;
+  }
+
+  throw new Error(`Path Can't be generated from collection as no matching category: ${metadata.category}.`);
 }
 
 function formatBucketName(bucketName: string): string {
-  if (bucketName.startsWith('s3://')) {
-    return bucketName.replace('s3://', '').replace('/', '');
-  }
+  if (bucketName.startsWith('s3://')) return bucketName.replace('s3://', '').replace('/', '');
   return bucketName;
 }
 
 /**
  * Generates specific dataset name based on metadata inputs
  *
- * @export
- * @param {string} region
- * @param {?string} [geographicDescription]
- * @returns {string}
+ * see {@link slugify} for how it is formatted
+ *
+ * @param region
+ * @param geographicDescription
+ * @returns
  */
 export function formatName(region: string, geographicDescription?: string): string {
-  if (geographicDescription) {
-    return slugify(geographicDescription);
-  }
+  if (geographicDescription) return slugify(geographicDescription);
   return slugify(region);
-}
-
-export function formatDate(collection: StacCollection): string {
-  const interval = collection.extent?.temporal?.interval?.[0];
-  const startYear = getPacificAucklandYear(interval[0]);
-  const endYear = getPacificAucklandYear(interval[1]);
-
-  if (startYear == null || endYear == null) {
-    throw new Error(`Missing datetime in interval: ${interval.join(', ')}`);
-  }
-  if (startYear === endYear) {
-    return startYear;
-  }
-  return `${startYear}-${endYear}`;
-}
-
-/**
- * Convert time zone-aware date/time string to Pacific/Auckland time zone string
- *
- * We can't convert the time zone of a `Date` directly, but instead have to produce a localised
- * date/time string. We arbitrarily convert to the New Zealand English format (For example,
- * '2/04/2024, 11:27:30 am'), since it doesn't seem to be possible to convert directly to a more
- * easily parsed format like ISO 8601 or RFC 3339.
- *
- * @param {string | null} dateTimeString Optional date/time string which can be parsed by the `Date` constructor
- * @returns {string | undefined} Localised date/time string
- *
- */
-function getPacificAucklandYear(dateTimeString: string | null): string | undefined {
-  if (dateTimeString == null) {
-    return undefined;
-  }
-
-  const pacificAucklandDateTimeString = new Date(dateTimeString).toLocaleString('en-NZ', {
-    timeZone: 'Pacific/Auckland',
-  });
-  return pacificAucklandDateTimeString.split(/[,/]/, 4)[2];
 }
 
 /*
@@ -184,33 +151,49 @@ function getPacificAucklandYear(dateTimeString: string | null): string | undefin
 /**
  * Gets tiff of first item listed in collection
  *
- * @async
- * @param {string} source
- * @param {StacCollection} collection
- * @returns {Promise<Tiff>}
+ * @param source
+ * @param collection
+ * @returns
  */
 export async function loadFirstTiff(source: string, collection: StacCollection): Promise<Tiff> {
   const itemLink = collection.links.find((f) => f.rel === 'item')?.href;
   if (itemLink == null) throw new Error(`No items in collection from ${source}.`);
+
   const itemPath = new URL(itemLink, source).href;
   const item = await fsa.readJson<StacItem>(itemPath);
   if (item == null) throw new Error(`Failed to get item.json from ${itemPath}.`);
+
   const tiffLink = item.assets['visual']?.href;
   if (tiffLink == null) throw new Error(`No tiff assets in Item: ${itemPath}`);
+
   const tiffPath = new URL(tiffLink, source).href;
   const tiff = await createTiff(tiffPath);
   if (tiff == null) throw new Error(`Failed to get tiff from ${tiffPath}.`);
   return tiff;
 }
 
+/**
+ * Load the ground sample distance from a tiff
+ *
+ * @throws if no GSD is defined
+ *
+ * @param tiff to load the data from
+ * @returns GSD if it exists
+ */
 export function extractGsd(tiff: Tiff): number {
   const gsd = tiff.images[0]?.resolution[0];
-  if (gsd == null) {
-    throw new Error(`Missing resolution tiff tag: ${tiff.source.url.href}`);
-  }
+  if (gsd == null) throw new Error(`Missing resolution tiff tag: ${tiff.source.url.href}`);
   return gsd;
 }
 
+/**
+ * Load the projection EPSG code from a tiff
+ *
+ * @throws if no ESPG code is defined
+ *
+ * @param tiff to load the data from
+ * @returns EPSG code if it exists
+ */
 export function extractEpsg(tiff: Tiff): number {
   const epsg = tiff.images[0]?.epsg;
   if (epsg == null) {